# internal/delivery/http/dto/user_response.go
package dto

type UserResponse struct {
	ID       string `json:"id"`
	Email    string `json:"email"`
	Name     string `json:"name"`
	Surname  string `json:"surname"`
	Phone    string `json:"phone"`
	Telegram string `json:"telegram"`
}